Allow setting/getting height/width of tabs
Add setters/getters for setting/getting the height/width of a tab.
If set to a non-zero value, this value will be used for the tab
height/width (margins/padding are in addition).  If set to zero,
then it will be automatically selected/calculated as before (line
height for top/bottom placement; width of 'A' for left/right).