fix(bar, slider): fix shadow drawing on short indicators
The issue is related to the case when the indicator is short but the main part (bg) clips it with a larger radius.
In this case this indicaotr might have a strange shape, like
 _
/ |
| |
\_|

We can't draw a shadow like this, so the indicator has to be et least a square,
to be sure it's not clipped to an ugly shape.

The previous code considered the bg radius to decide whether to draw shadow on short indicators or not.
However if the indicator's radius should be considered, becuse if its area at least a square
we can already draw shadows.