lvgl: kscan: set touch points within range
Check whether input points are outside of display and set to the maximum of the its axis.
Currently when an input device sends a point out of range, lvgl reports that it's outside of it.
That's a workaround avoiding implementation of input device calibration.
Maybe a calibration functionality could be implemented within kscan, rather than doing for each different input device driver?