zephyr: demote error message to debug level
On some I2C controllers, when using gestures, we can get quite a number
of spurious events. Don't pollute the log unnecessarily especially as
this doesn't impact the functionality. Demote the error log to debug when
this happens and also correct the typo in the message itself (it should
say it: "Could *not* ...".

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@huawei.com>