zephyr: Inhibit some partial EPD display updates
LVGL sometimes needs to render the screen in multiple chunks. This
causes excessive flickering on EPD displays. The EPD drivers (SSD16xx
and GD7965) in Zephyr can inhibit updates by using the blanking setting.

Update the glue code to inhibit blanking when flushing multiple chunks
to an EPD display. When the last chunk has been written to the
display, disable blanking and force an update.

Updates that consist of a single chunk will be written without first
blanking the display. This will trigger a partial update on displays
that support it.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>