zephyr: Pick up display from /chosen/zephyr,display
The LVGL binding currently uses a device name from Kconfig to discover
the display. This is fragile and won't work well when removing labels.

Switch to using the /chosen/zephyr,display property instead. This
ensures that there is a compile-time error instead of a runtime error
if the display can't be found.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>