zephyr: zephyr: Pick up keyboard scan from /chosen/zephyr,keyboard-scan
The LVGL binding currently uses a device name from Kconfig to discover
the kscan device. This is fragile and won't work well when removing labels.

Switch to using the /chosen/zephyr,keyboard-scan property instead. This
ensures that there is a compile-time error instead of a runtime error
if the keyboard scan device can't be found.

Signed-off-by: Kumar Gala <galak@kernel.org>