feat(vglite) Cleanup vglite.
1. Remove the internal blit structure passed as refference.
Even though now the APIs have more arguments, this will make more
clear what is the function purpose and what it needs.
Also it will separate the normal blit and blit with transformation.
More cleanup will be needed to have tinner functions and faster for simple tasks.
2. Rename some booleans to be more suggestive.
3. Rename _lv_gpu_nxp_vglite_check_blit() to check_src_alignment() and remove unneeded checks.
4. Cleaup a bit the split_blit.
Remove the functions to draw rectangle around tiles (can use vglite rect API instead.
Remove the sw_blit() fallback, the error shall be propagated to draw ctx.
5. Allow the transformation to use split_blit. It might be enough the first stage:
"moving starting pointers as close as possible to [x1, y1], so coordinates are as small as possible."
More cleanup is needed on this area.
6. Use LV_LOG instead of LV_LOG_ERROR for VG_LITE_LOG_TRACE.

Signed-off-by: Nicușor Cîțu <nicusor.citu@nxp.com>