feat(vglite) Move the vgbuff helpers to a separate file.
Dest vgbuff will be initialized just once at the beginning of the frame
for all blit/draw operations.
Src vgbuff will be initialized for each blit but only if the buffer
address is different. E.G. If you blit the same image multiple times,
it will be set just once.

Split blit workaround:
- needs to set only the new buffer address because it has been shifted
(to have the coordinates as small as possible)
- all the other vgbuff attributes will stay in place (width/height/stride...)

Signed-off-by: Nicușor Cîțu <nicusor.citu@nxp.com>