fix(vglite) Fix blit with transformation
Normal blit: blend_area is the intersection of coords and clip_area.
Scissor shall be set to blend_area as clip_area might be a bit bigger.

Blit with transformation: blend_area is the coords.
The transformation matrix will take care of rotation/zoom and set the proper destination area.
Scissor shall be set to clip_area in this case.

Signed-off-by: Nicușor Cîțu <nicusor.citu@nxp.com>