feat(meter): Add linear meter widget
Based on the current meter expand them for supporting vertical
and horizontal linear meter. The algorithms and api for
the old meter is still the same for the backward compatibility.
For using another form of meter just change the form(by default form is round):

 `lv_meter_set_indicator_form(obj, LV_METER_INDICATOR_FORM_LINEAR);`

Also you can set the direction of scale(by default: left to rigth and up to down):

 `lv_obj_set_style_base_dir(meter, LV_BASE_DIR_RTL, 0);`

And swap the labels with needle, if needed:

 `lv_meter_set_needle_side(scale, toggle_side);`

If widget width >= height the result form is horizontal otherwise it is a vertical.
All current meters can be linear, EXCEPT the clock meter, because you
know, it's a clock :)

issue #10986