still tweaking the reST files.
I have come up with a sort of solution for the API sections. I am using the `:ref:` primer to create a link. I am using the same primer for linking to other sections in the documentation. I have updated some of the links in the files but not all of them yet. I believe I am done adding the sphinx directives and roles to the reST files.

I also changed the maximum width the documentation is allowed to be shown at. I thought that 800px wide was a little on the narrow side considering most computer displays are 1920px wide. if going full screen it leaves a pretty sizeable bar on the right hand side that is empty.

This commit builds with no errors or warnings :thumbsup: