zephyr: enable setting full_refresh bit for LVGL displays
Enable setting full_refresh bit for LVGL displays, which forces LVGL to
refresh the entire framebuffer on each call to display_write.

This is useful for display hardware that can directly render a framebuffer to
the display, as the LVGL buffer can now be passed directly to the
display hardware without expensive memory copies into an intermediate buffer.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>