test(dropdown): add test to ensure get_option_index checks string length
Previously there was a bug where lv_dropdown_get_option_index would only
check the first n chars when looking up an option (where n is the
shorter of the option length and the item length). This would cause "ab"
to test equal to "abc" because it only checked the first 2 chars. This
bug is fixed now, and this test will prevent it from being reintroducted
inadvertently.