changes a couple of things to get the CPython binding working.
These are pretty minor changes. One is removing the declaration for the `_lv_font_clean_up_fmt_txt` function. The function wasn't defined which caused linking problems.

I also moved a couple of structures that were defined in source files and put them into the header files. If this is an issue let me know, I will see if I can manage to work around them not being exposed. I am not sure if ctypes will play nice with me if it is unable to determine the size of the structures. Because ctypes is disconnected so to speak from the structures in C code The only way for me to show their existence is if they can be seen in the header files. That includes the fields in the structures. This is if the structure is used aas a type for aa field in a public structure or as a type in a public function.

They are small changes, I hope it's not a big deal to make them. I managed to code around the rest of the issues I was having before.

The code generation script is too specific to LVGL for my liking, so it will need maintenance due to it having specific parts of it coded specifically to handle LVGL, if those things change it could break. Not a huge deal.