feat(vglite) Add VGLITE_TASK_BUF support.
Add vglite_cmd_buf_is_flushed() to get the flush status of the GPU. If the GPU is busy, the cmd buffer will not be flushed but will be appended by one more task (or command).

Add static circular buffer to hold the pending tasks (queued and flushed). Two indexes, head and tail, are used to signal the beginning and the end of the valid tasks that are pending.
VGLite thread has to run as much as there are pending tasks in the buffer. Every time the GPU cmd buffer has been flushed, we search through the task buffer and update the status:
	- the previously flashed tasks are signaled as ready to dispatcher.
	- the queued tasks are transitioning to flushed.
The above process runs only in multithreading mode (LV_USE_OS). In v8 compatibility mode, GPU process the tasks and signal them as complete one by one.

Signed-off-by: Nicușor Cîțu <nicusor.citu@nxp.com>