feat(freertos) Define USE_FREERTOS_TASK_NOTIFY.
A direct to task notification is an event sent directly to a task, rather than
indirectly to a task via an intermediary object such as a queue, event group or semaphore.
Sending a direct to task notification to a task sets the state of the target task
notification to 'pending'. Just as a task can block on an intermediary object such
as a semaphore to wait for that semaphore to be available, a task can block on a task
notification to wait for that notification's state to become pending.

Unblocking an RTOS task with a direct notification is 45% faster and uses less RAM
than unblocking a task using an intermediary object such as a binary semaphore.

RTOS task notifications can only be used when there is only one task that can be the recipient of the event.

Signed-off-by: Nicușor Cîțu <nicusor.citu@nxp.com>