Refresh test code.
I changed the organization of the code so double buffering will actually work. I also made some changes to the performance monitor so a true FPS fill be given, it will not count frames when there is no work to be done.

I have only tested this on my desktop PC with the default refresh timer set to 1 millisecond Test version and also baaseline version has the changed made to the performance monitor so they would not count frames where no work has been done. This gives a true indication of what kind of difference there would be performance wise between the original refresh setup and the one I modified. Using a slider and having it change value every 1 millisecond using a frame buffer that is the size of the display (480 x 320 x 32) the code change I made gave a boost of 20% which is a pretty large increase.

I would like someone to test it in a read world setup on an MCU using DMA memory and double buffering. I believe there is going to be an even larger increase in performance when this is done. The default refresh timer value for this test should be decreased to 1 millisecond to ensure the timer value is not limiting the speed in which an update is able to occur and the slider value should only be changed once a millisecond.