build(cmake): pass freetype build flags to parent
If using freetype, the parent CMake project will need to know the
freetype include and link dirs as well as the list of freetype libraries
to link to.

This commit defines an option in the CMake project - LV_USE_FREETYPE -,
which, if set, makes LVGL look for the freetype library using pkgconfig,
and add the include dirs, link dirs and libraries to the cmake library,
so that these are automatically passed along to dependants through the
PUBLIC interface.

In the example below:
-------
add_subdirectory( path/to/lvgl )
add_executable( myapp ... )
target_link_libraries( myapp PRIVATE lvgl:lvgl )
-------

If lv_conf.h has:

The build will fail because the compiler will need to know about the
includes and freetype libraries.

This commit fixes that and spares parent projects from having to add
the freetype library by hand.

Signed-off-by: António Oliveira <antonio@amsobr.com>