Adds additional functions to change the way the sdl window appears.
Adds functions for setting position, brightness and other various things for an SDL window. It also exposes flags for controlling various elements of the the window creation, like resizing and displaying a title bar and border.

This is an API Breaking change. There is a new parameter added to the create function. If this parameter is supplied a zero it will behave exactly like it did prior. If a -1 is supplied to the additional parameter that would be specifying no flags at all. The flags can be or'ed together.

The flags are in a new enumeration named  `lv_sdl_window_flags_t`. There is a flag that is added to center the window to the desktop and that is used as a default.

There is also a flag for high dpi support if the OS and video driver support it.