Update lv_example_roller_1.c - lv_roller_set_options() parameter issue
The example uses the old format for passing options to the lv_roller_set_options() function call, which results in a complile error. 

The "Styling the roller" example which is the next one to follow in the documentation does it correctly.

As a result, the attached code has been modified to use the new format for passing options.