fix(tiny_ttf): fix font metrics confusion
Fix underline position. It is relative to the baseline.

Change lv_tiny_ttf_create_xxx() and lv_tiny_ttf_set_size() to accept
font size rather than line height. Font size is the more conventional
way to specify font scale. It is also more consistent with the naming of the
function, and the names of the built-in fonts in lvgl.