Update lv_color.h with HIGH_SPEED_COLOR_MIX_RGB565 routine
Routine is highly optimized to each case of 0-32 possible cases for blending value (mix+4)>>3 and doesn't use multiplication at all.
It is written to be "assembler friendly". It is thoroughly tested for correct calculation and preservation of linearity in mixed result.

Speed testing should be performed on different platforms. Maybe, it can be included in benchmarks and measure how does each benchmark performs with turned ON/OFF this routine on systems with RGB565 format.