feat(draw_buf): add ALLOCATED image header flag
For image decoder, the decoded image could be the original image which may stay in flash, thus no need to free.
For canvas, a static draw buf could be used, thus also no need to free.
Add the ALLOCATED flag can make it easy to distinguish them. While the existing MODIFIABLE is only used for modifing image in place

Signed-off-by: Xu Xingliang <xuxingliang@xiaomi.com>