Fixes more warnings.
There are warnings left I am waiting for a reply concerning the renderers. I also have to do some research into the "WARNING: Unparseable C++ expression" errors. I know why the errors are occurring I just need to find a solution other than removing some of the stuff in the expressions. I am probably going to have to make a custom directive to handle the additional elements in the expressions.