fixes incorrect enum to integer lv typedef
enumerations by default are integers and in LVGL they are retyped to other interger types like unsigned. This change makes a correct connection between an enumeration and the lvgl type that is supposed to be used for it.

I have also added 2 additional json fields to the macro json_type. These fields allow the user to create representations of the macros in whatever language they are making a binding for. The 2 new fields are 'params' and 'initializer'. params will be NULL if the macro is a constant and takes no arguments otherwise it will be a list of parameter names. initializer if the value or contents of the macro. This could be code that is to be executed or it could be a simple integer value. The user will have to figure that out and how to parse the data.