fix(cmake): generate versioned shared libraries
Add missing version suffix to shared libraries. Currently the filename of
generated shared libraries is only liblvgl.so, which prevents coexistence
of different versions of LVGL on the same system. Set VERSION and SOVERSION
to make cmake add the version suffix to generated shared libraries. That
changes the filename to liblvgl.so.9.1.1 and includes symlink with major
ABI version, i.e. liblvgl.so.9 .

Signed-off-by: Marek Vasut <marex@denx.de>