Account for the frame buffer virtual resolution deviating from the visible resolution.
The Linux frame buffer flush_cb method now accounts for the virtual resolution deviating from the visible resolution. The offset into the paint buffer is calculated based on the visible resolution, and the offset into the frame buffer is calculated using the line_length, i.e. stride. The x offset and y offset are only used to calculate the position into the frame buffer, not the position into the paint buffer.