fix(vg_lite): fix incorrect cache operation
After the CPU decodes the image, it should call the flush D-Cache operation
to completely write the pixel data in the cache to RAM.

The invalidate D-Cache operation is incorrect, which will cause the data
in the cache that is not written to RAM to be lost.

Invalidate cache is usually done when the peripheral writes to the memory
first and the CPU needs to read it.

Signed-off-by: pengyiqiang <pengyiqiang@xiaomi.com>