Fixes collecting the version from the LVGL source files.
I had also noticed there was some version differences in different areas of LVGL. So I write a python script that will change all of the version numbers in LVGL all at the same time.

The script gets run like this.
```shell
scripts/set_version.py 9.1.1.dev
```

Even if there isn't anything to set in the last space you need to still supply the dot with nothing after it.

```shell
scripts/set_version.py 9.1.1.
```

This will change the version located in the following files

* library.json
* library.properties
* lv_version.h
* Kconfig

Let me know if there are any that I missed and I will add them.
I have pointed everything I was able to locate to using the find_version.py script.

The CMAKE script has been updated to use this file and to do what it needs to do with the version number that has been collected.

The document build system has also been fixed, it will run and complete if there is a documentation error. A documentation error is something missing. If there is a catastrophic error in Doxygen or in Sphinx that will still cause the build to fail.