reworks the display refresh callback.
There was no reason to have the wait for the buffer flushing called more than a single time during the refresh. I move it so it is being called right away.

I also removed the else statement in the wait_for_flushing function that would do the spinning wheels and the automatic setting of disp-> flushing to zero if the the flush_wait_cb has been set. The user is able to call the flush ready function from inside of the callback should they want to notify LVGL that the buffer has actually flushed. That is the reason why I removed the else statement that checks to see if the `disp->flushing` flag has been set. This allows the user to use the callback to complete other tasks while a buffer is flushing.

It leaves it up to the user to decide if that callback is to be used as a way to stall LVGL without having a spinning wheels while the buffer is transferring or if they want to use that callback to do something else like checking pin states on an MCU in the event there is a wait that has to take place.

The other thing I added was an exit from doing the refresh if using double buffering instead of sitting there spinning it's wheels. for the most part double buffering and DMA memory use go hand in hand. The benefit it to be able to do other things while the buffer is still being transmitted. LVGL causing a stall is wasted processor time. So when is done is the timer duration gets reduced to a 1 so next time lv_task_handler gets called the refresh will check if the buffer has finished and if it has then it will perform the refresh. This allows other things to get processed instead of wasting processor time. This can be enabled or disabled using the LV_RETURN_FROM_FLUSH_WAIT which is in the config template.

I also changed the refr_now function so it will inform the user if the refresh has completed or not. It will only return false if double buffer is enabled and LV_RETURN_FROM_FLUSH_WAIT is set to 1 and the refresh doesn't complete. Otherwise it will always return true