fix(kconfig): Remove closing square bracket from LV_TXT_BREAK_CHARS
This removes the closing square bracket from the default value of
LV_TXT_BREAK_CHARS. This is done because the zephyr build tool `west`
relies on the python package `kconfiglib` for parsing Kconfig files. It
seems like `kconfiglib` has a bug when handling this character within a
string default value. Unfortunately the package is unmaintained so
issuing a fix and incorporating into `west` is not possible.