fix(decoder): replace portable (C99) with conventional format specs
The format spec "%d" is not compatible with the actual declaration of
these variables (uint32_t), but they seem to be getting promoted to the
compatible type int in the enclosing function.

This causes a compile error (-Werror=format=) when using the current
spec (LV_)PRId32 — defined as unsigned long (ESP-IDF, ESP32, ISO C17).