fix:  restore missing examples
Why needed:  these examples are missing in on-line documentation:
- button       (renamed)
- buttonmatrix (renamed)
- colorwheel   (removed)
- image        (renamed)
- imagebutton  (renamed)
- meter        (removed

This script that generates the examples pages evidently was overlooked.
Also, it appears that all the plumbing code for `lottie` examples is
present, but was not added, so I added it.  I need to confirm that those
examples are actually ready for publication.