fix: docs error blocking API generation under Windows...
...in the case where TMPDIR environment variable pointed to a drive
different than the LVGL project directory.  (TMPDIR is used by Python
`tempfile.mkdtemp()` used in `build.py`.)  The result was on Windows
Doxygen was running in the LVGL project's `./docs/` directory instead
of the temporary directory where it was intended to run, causing
Doxygen to generate XML files for the API pages in in the wrong place.
The solution was to replace the attempted `cd` with a cross-platform
compatible `cd` built into Python's `os` library by temporarily having
the `build.py` script change directories to where Doxygen should run
and change back again after running it.

Additionally:

1.  Handled 2 Python warnings where a regex string was using an illegal
    escape sequence by making it a "raw" string, since the escape
    backslash needs to be passed un-interpreted to the regex.

2.  Replaced `build.py`s get_version() function with one that works
    on both Linux and Windows platforms.  Credit:  @kdschlosser.

Resolves #6924.