fix(bench) do not average performance with non-rendered frames
If frames are not rendered the FPS and CPU utilization averages are
calculated with what appears to be zero effort from the device. The
benchmarks are skewed towards unrealistic high FPS and low CPU
utilization even though the scenes are very slow in real life.

This can happen when there are few redraws on scenes that change
infrequently.

Signed-off-by: Cristian Stoica <cristianmarian.stoica@nxp.com>