fix(buttonmatrix):  fix 1-off error in `lv_buttonmatrix_get_button_text()`
lv_buttonmatrix_get_button_text() function has this code in it to avoid
attempting to use a user-provided btn_id value that will make the
`while(btn_i != btn_id)` loop proceed past the end of the array of buttons.
However, it has a 1-off error which does permit it to go 1 past the end
of the array.  This PR fixes that error.

Resolves #7192