fix(docs):  fix bad highlight color
In Dark Mode in the documentation, the color that is used to highlight
recognized function names is so dark that it is nearly invisible:  it
nearly blends into the black background.  In LIGHT mode, it is so dark
it is nearly black against the light background so I almost cannot tell
it is highlighted.  This color update fixes both.

Resolves #7195