fix(chart):  `lv_malloc()` was allocating double size needed
When a chart is a SCATTER type, each series must have both an array
of X values in addition to Y values (both having `chart->point_cnt`
`int32_t` elements).  This call of `lv_malloc()`, occurring when the chart type
was changed to SCATTER type, was allocating a space for an `lv_point_t`
(2 `int32_t` values) instead of 1, for each X value -- double the amount
actually needed.  No harm was done to the application except for
wasting RAM.

No issue was created for this.