fix(draw):  fix bug introduced by PR #6638
PR #6638 introduced a bug that caused issue #7230.  Since
I did not see a fix already submitted for it, I am submitting it
here.  This fix is tested under `lvgl/lv_port_pc_visual_studio`
and it resolves both the issue seen with the Knob as well as
with the Roller Widget display updates after the Widget's
value has been changed.

Resolves #7230