feat(docs):  batch 7 of proofread/edited docs
- slider.rst
- span.rst => spangroup.rst
- spinbox.rst
- spinner.rst
- switch.rst
- table.rst

Most of these edits are to correct grammar and clarify meaning.
Spangroup got an extra paragraph in its Overview section for
reader-orientation purposes.

span.rst was renamed (along with its title).  A redirect HTML file
is created in place of the old `span.html` in case anyone linked to
it in the field.