feat(anim):  clarify reverse play in animation API
The term "playback" is used in anim.c/.h to mean "reverse play".
Unfortunately, these are opposites in the English language and so to
safely use animations, a native-English programmer must study the C code
to figure out what it really means.  (Documentation did not previously
clarify the meaning.)  This PR clarifies this meaning without breaking
the API.  (API "name patches" are added into `lv_api_map_v9_1.h` to
preserve [not break] calls for code in the field.)

Note:  major edits to animation.rst document is currently pending in
PR #7324 so its update to match (with changed function names) will need
to come separately after it is committed to prevent difficult conflicts.

Resolves #7331