feat(docs):  batch 10 of proofread/edited docs
- color.rst
- draw.rst
- index.rst (under `/docs/details/widgets/`)

`draw.rst` is another big one that required quite a bit of diggin in the
C code to sort out what some things meant.

`index.rst` refers to `spangroup` which changed names and somehow got
out of sync with what batches of proofread/edited documents are pending
approval and being committed.  So I changed it again here so that the
link to `spangroup.rst` would be intact for the Widgets overview page.