feat(docs):  add clarity of meaning to Length Units section
In the discussion of documenting DPX in PR #7306, I realized that the
way it was currently documented (in `coord.rst` under section merely
called "Units", that the readers in the discussion were unaware of
`percent` and `pixels` having already been documented in the first major
section of this document.  Thinking about this, I realized that the
existing "Units" section does not convey well to the reader that it is
a HUGE PART of LVGL Styles and applies across the boards to every
part of the drawing domain when a length between 2 points is calculated,
which is ubiquitous across virtually all of the drawing logic.

So I made it a point of calling attention to the actual breadth and depth
of the meaning of this section by highlighting its real meaning.