fix(doc_builder.py):  fix bugs preventing running...
...on Windows platform where the temporary directory directory generated
by Python's `tempfile.mkdtemp()` is on a different drive.  When it is,
it causes  the "cd tmpdir && doxygen Doxyfile" syntax to fail because of
the different semantics of the `cd` command on Windows:  it doesn't
change the default DRIVE if `cd` is executed from a different drive.
The result, when this is the case, is that Doxygen runs in the current
working directory instead of in the temporary directory as was intended.

Also:  LVGL_URLPATH environment variable was being populated with a
multi-line list of all current LVGL branches, rather than just the name
of the current branch as intended.

The result:  the `gen_json.py` script now runs correctly on a Windows
platform with TMPDIR environment variable set to a different drive
than the script is executing from.