adds option to gen_json to build without docstrings
Currently, in the MicroPython binding gen_json.py is used to properly map a callbacks userdata that is stored in a nonpublic structure. While doing this works without any issues it takes a bit of additional time for the gen_json.py script to run. This additional time is caused by the parsing of the docstrings from the header files. The docstrings do not need to be parsed in order to provide the information that the MicroPython binding code generator needs in order to work. Having the ability to turn off the docstrings will speed up the compile time, it also removes doxygen as a dependency to compile the binding.

I have also removed the threading from the gen_json.py script that was used to output dots to stdout. Having this causes issues if inporting the gen_json.py script instead of running is using a subprocess.