fix(docs):  clarify obj hierarchy after `lv_menu_page_create()`
In Issue #7600, @phelter correctly pointed out that the documentation
for `lv_menu_page_create()` (and its 1st argument name) is misleading
about the resulting object hierarchy after the call.  This submission
both clarifies the result in the API documentation as well as adds some
enhancements to `menu.rst`, though it still needs a thorough review.

Resolves #7600