docs:  fix doc-build failure...
We moved the ./docs/_static/ directory to ./docs/src/_static/
out of necessity so that the `sphinx-build` would work since it
uses the _static/ directory.  I suspect what is being built by
the CMAKE (emscripted stuff) MAY need to live in its own directory.

What I do know is where the build is currently failing is the
`cp` command at the end of the ./scripts/build_html_examples.sh
script.

What I don't know if there might be something down the line
after this that will fail because it is expecting the `built_lv_examples`
to be where it used to be....

So instead of trying to move that, I have created an empty file
in ./docs/_static/README.md

So that the MERGE will have that directory present.  We can
study where the DEPLOY step is wanting to get this from later.

As I wrote in the README.md, I suspect that the Sphinx's _static/
directory shoult NOT have been used for this, but rather the
`built_lv_examples` should have its own directory for deployment.
But that's for a later task.