docs(requirements.txt):  remove jquery requirement
Turned out to not be needed because it places inclusion of jQuery
at the tail end of the HTML page, but it is needed at the top so that
functions with jQuery calls in them can be compiled correctly as the
HTML page is being loaded.  I believe this is a bug in the extension,
so we won't be using it for now and instead are hard-coding it in
the `page.html` template.