docs(introduction.rst):  fix bug in API links...
The symbol `lv_display_flush_ready` was being seen and tested for in the
code symbol tables generated by Doxygen, but as it turns out, the
prototype (where the documentation is for that function) is, as well
as the function itself in the `.c` file, is prefixed with the symbol
`LV_ATTRIBUTE_FLUSH_READY`.  The code looks like this:

    /**
     * Call from the display driver when the flushing is finished
     * @param disp      pointer to display whose `flush_cb` was called
     */
    LV_ATTRIBUTE_FLUSH_READY void lv_display_flush_ready(lv_display_t * disp);

Unfortunately, this is tricking Doxygen into not being able to parse that
line and so the documentation that goes with it never makes it into the
Doxgen XML output.

The result was that the `lv_display_flush_ready` was not found in the
symbol table, and that was causing the `lv_display.h` API link to be
omitted from the bottom of the document!

Since this prefix appears to be vital (e.g. probably can be to be 'inline'
or some other thing [perhaps connected to an interrupt?] for certain
build environments), so the solution was simply to choose ANOTHER
symbol from the `lv_display.h` file and replace `lv_display_flush_ready`
with it to get the API-page link to `lv_display.h` to be added to the
document.  The symbol chosen was `lv_display_flush_wait_cb_t`,
and it is working now and reflected at the end of the on-line document
[here](http://crystal-clear-research.com/for_gabor/demo/docs.lvgl.io/master/intro/introduction.html)
(in the API section at the end).